requirements.txt
Django
awesome-slugify
